Split runtime_checks.py: solution/validate_V.py + lcm/_transition_checks.py
regime_building/ is the build-time pipeline (UserRegime → engine.Regime).
The runtime checks fire on solve / simulate, not at build, so they don't
belong here. Split by caller:

- validate_V (+ its diagnostic-attachment helpers) is a tight subroutine
  of the backward-induction loop in solve_brute.py and the V handed to
  simulate.py. Moves to solution/validate_V.py (solve owns producing V;
  simulation imports from there).
- The regime-transition-probability pre-flight sweep is called from
  Model.solve()/simulate() before backward induction runs. Moves to a
  top-level lcm/_transition_checks.py — orthogonal to solution/ and
  simulation/, sibling of engine.py and model_processing.py.

Imports updated in model.py, solve_brute.py, simulate.py and three tests.
Docstring on user_regime.validate_transition_probs updated to point at the
new location.
