+++ tests/test_beartype_claw.py
@@ -25,6 +25,7 @@ import pytest
 from beartype.roar import BeartypeCallHintViolation
 
 from lcm import AgeGrid, LinSpacedGrid, Model
+from lcm._transition_checks import _validate_regime_transition_probs
 from lcm.exceptions import (
     GridInitializationError,
     ModelInitializationError,
@@ -32,7 +33,6 @@ from lcm.exceptions import (
 )
 from lcm.interfaces import _build_regime_sharding
 from lcm.model import _validate_log_args
-from lcm.regime_building.runtime_checks import _validate_regime_transition_probs
 from lcm.simulation.simulate import _compute_starting_periods
 from lcm.solution.solve_brute import _log_per_period_stats
 from lcm.state_action_space import _validate_all_states_present
@@ -61,8 +61,8 @@ def test_claw_checks_lcm_solution() -> None:
         )
 
 
-def test_claw_checks_lcm_regime_building_runtime_checks() -> None:
-    """Type-violating arguments to runtime_checks helpers raise."""
+def test_claw_checks_lcm_transition_checks() -> None:
+    """Type-violating arguments to `lcm._transition_checks` helpers raise."""
     with pytest.raises(BeartypeCallHintViolation):
         _validate_regime_transition_probs(
             regime_transition_probs={"working": jnp.array([1.0])},  # ty: ignore[invalid-argument-type]
